Handle multiple games of the same player during one round
If a player plays multiple games during a round, treat these as played
simultaneously. That is, the rating adjustments are calculated based on
the ratings before the round and added together. This is the same as the
EGF ratings handle games played in one tournament.

Analogously to EGF rules for tournaments, do not decrease a player's
rating by more than 100 points in a single round.