Allow scheduling a game manually
Allow adding a game to a round manually, by specifying the players and
optionally handicap and/or result. Unlike automatic scheduling, this
allows having a player play multiple games in one round and scheduling
an absent player.

If the handicap is not specified, it is calculated from the current
ratings (which includes any game results already entered for this
round). If black is stronger than white, the handicap is always 0.

To enter multiple custom games, input one at a time and click Apply
changes for every game.